svg change Integer Properties to Double

+++ svg/src/main/java/jetbrains/jetpad/projectional/svg/SvgEllipse.java
@@ -5,9 +5,9 @@ import jetbrains.jetpad.model.property.ValueProperty;
 import jetbrains.jetpad.values.Color;
 
 public class SvgEllipse extends SvgElement {
-  public final Property<Integer> cx = new ValueProperty<>(0);
-  public final Property<Integer> cy = new ValueProperty<>(0);
-  public final Property<Integer> rx = new ValueProperty<>(10);
-  public final Property<Integer> ry = new ValueProperty<>(20);
+  public final Property<Double> cx = new ValueProperty<>(0.0);
+  public final Property<Double> cy = new ValueProperty<>(0.0);
+  public final Property<Double> rx = new ValueProperty<>(10.0);
+  public final Property<Double> ry = new ValueProperty<>(20.0);
   public final Property<Color> fill = new ValueProperty<>(Color.BLACK);
 }
